# gen/http/openapi.yaml
swagger: "2.0"
info:
    title: Orbo Video Alarm System
    description: Open source video alarm system with motion detection and Telegram integration
    version: "1.0"
host: localhost:8080
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
...
    /api/v1/system/detection/start:
        post:
            tags:
                - system
            summary: start_detection system
            description: Start motion detection on all active cameras
            operationId: system#start_detection
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/SystemStartDetectionResponseBody'
                        required:
                            - cameras
                            - motion_detection_active
                            - notifications_active
                            - uptime_seconds
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/SystemStartDetectionInternalResponseBody'
                        required:
                            - message
            schemes:
                - http
...
definitions:
...
    CameraInfoResponseBody:
        title: CameraInfoResponseBody
        type: object
        properties:
            alerts_enabled:
                type: boolean
                description: When false, detection pipeline still runs for bounding boxes but no events are created or alerts sent.
                example: false
            created_at:
                type: string
                description: Creation timestamp
                example: "1971-11-22T08:53:09Z"
                format: date-time
            device:
                type: string
                description: Camera device path (e.g., /dev/video0)
                example: Praesentium debitis aperiam.
            fps:
                type: integer
                description: Frames per second
                example: 2017387124810930427
                format: int64
            id:
                type: string
                description: Camera unique identifier
                example: ecde416f-f010-11f0-998b-5847ca7b8ce1
                format: uuid
            name:
                type: string
                description: Camera name
                example: Rerum molestiae voluptatem est quia nulla eligendi.
            resolution:
                type: string
                description: Camera resolution
                example: Quas dolore culpa est dignissimos.
            status:
                type: string
                description: Camera status
                example: active
                enum:
                    - active
                    - inactive
                    - error
        description: Camera information
        example:
            alerts_enabled: false
            created_at: "2013-11-17T08:21:24Z"
            device: Neque commodi ea fuga veniam qui.
            fps: 72305995745697673
            id: ecde43b5-f010-11f0-998b-5847ca7b8ce1
            name: Ut odio voluptatem tempore sit.
            resolution: Et earum earum quis voluptatum et ut.
            status: active
        required:
            - id
            - name
            - device
            - status
...
    SystemStartDetectionInternalResponseBody:
        title: SystemStartDetectionInternalResponseBody
        type: object
        properties:
            message:
                type: string
                description: Error message
                example: Quas est sit aut.
        description: Failed to start motion detection
        example:
            message: Quis reprehenderit aut rerum.
        required:
            - message
    SystemStartDetectionResponseBody:
        title: SystemStartDetectionResponseBody
        type: object
        properties:
            cameras:
                type: array
                items:
                    $ref: '#/definitions/CameraInfoResponseBody'
                description: Camera status list
                example:
                    - alerts_enabled: false
                      created_at: "2005-10-26T18:36:02Z"
                      device: Explicabo earum.
                      fps: 3590518115910755142
                      id: ecdd4571-f010-11f0-998b-5847ca7b8ce1
                      name: Et est corporis porro illum.
                      resolution: Velit ullam et dolores sequi temporibus eum.
                      status: error
                    - alerts_enabled: false
                      created_at: "2005-10-26T18:36:02Z"
                      device: Explicabo earum.
                      fps: 3590518115910755142
                      id: ecdd4571-f010-11f0-998b-5847ca7b8ce1
                      name: Et est corporis porro illum.
                      resolution: Velit ullam et dolores sequi temporibus eum.
                      status: error
            detecting_cameras:
                type: integer
                description: Number of cameras currently running detection
                example: 1701655547543221956
                format: int64
            motion_detection_active:
                type: boolean
                description: Motion detection status
                example: true
            notifications_active:
                type: boolean
                description: Notification status
                example: false
            pipeline_detection_enabled:
                type: boolean
                description: Whether AI detection is enabled (false when mode=disabled)
                example: true
            pipeline_detectors:
                type: array
                items:
                    type: string
                    example: Et commodi sapiente dignissimos quidem.
                description: Enabled detectors in the pipeline
                example:
                    - Reprehenderit ut.
                    - Et velit.
                    - Explicabo eligendi qui corporis repudiandae deleniti.
            pipeline_execution_mode:
                type: string
                description: Pipeline execution mode (always sequential)
                example: Porro maxime et repudiandae aut quae dolore.
            pipeline_mode:
                type: string
                description: Current pipeline detection mode (disabled, continuous, motion_triggered, scheduled, hybrid)
                example: Suscipit sunt voluptas officiis eum.
            uptime_seconds:
                type: integer
                description: System uptime in seconds
                example: 1863459564847017821
                format: int64
        example:
            cameras:
                - alerts_enabled: false
                  created_at: "2005-10-26T18:36:02Z"
                  device: Explicabo earum.
                  fps: 3590518115910755142
                  id: ecdd4571-f010-11f0-998b-5847ca7b8ce1
                  name: Et est corporis porro illum.
                  resolution: Velit ullam et dolores sequi temporibus eum.
                  status: error
                - alerts_enabled: false
                  created_at: "2005-10-26T18:36:02Z"
                  device: Explicabo earum.
                  fps: 3590518115910755142
                  id: ecdd4571-f010-11f0-998b-5847ca7b8ce1
                  name: Et est corporis porro illum.
                  resolution: Velit ullam et dolores sequi temporibus eum.
                  status: error
            detecting_cameras: 8020092528824024152
            motion_detection_active: true
            notifications_active: true
            pipeline_detection_enabled: false
            pipeline_detectors:
                - Reprehenderit iure.
                - Delectus itaque sed deleniti quas.
            pipeline_execution_mode: At labore enim.
            pipeline_mode: Fugit ipsam qui sint incidunt assumenda unde.
            uptime_seconds: 2100797482202161546
        required:
            - cameras
            - motion_detection_active
            - notifications_active
            - uptime_seconds
...