# gen/http/system/server/encode_decode.go
// Code generated by goa v3.14.1, DO NOT EDIT.
//
// system HTTP server encoders and decoders
//
// Command:
// $ goa gen orbo/design

package server

import (
	"context"
	"errors"
	"net/http"
	system "orbo/gen/system"

	goahttp "goa.design/goa/v3/http"
	goa "goa.design/goa/v3/pkg"
)

// EncodeStatusResponse returns an encoder for responses returned by the system
// status endpoint.
func EncodeStatusResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*system.SystemStatus)
		enc := encoder(ctx, w)
		body := NewStatusResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// EncodeStartDetectionResponse returns an encoder for responses returned by
// the system start_detection endpoint.
func EncodeStartDetectionResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*system.SystemStatus)
		enc := encoder(ctx, w)
		body := NewStartDetectionResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// EncodeStartDetectionError returns an encoder for errors returned by the
// start_detection system endpoint.
func EncodeStartDetectionError(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder, formatter func(ctx context.Context, err error) goahttp.Statuser) func(context.Context, http.ResponseWriter, error) error {
	encodeError := goahttp.ErrorEncoder(encoder, formatter)
	return func(ctx context.Context, w http.ResponseWriter, v error) error {
		var en goa.GoaErrorNamer
		if !errors.As(v, &en) {
			return encodeError(ctx, w, v)
		}
		switch en.GoaErrorName() {
		case "internal":
			var res *system.InternalError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			var body any
			if formatter != nil {
				body = formatter(ctx, res)
			} else {
				body = NewStartDetectionInternalResponseBody(res)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusInternalServerError)
			return enc.Encode(body)
		default:
			return encodeError(ctx, w, v)
		}
	}
}

// EncodeStopDetectionResponse returns an encoder for responses returned by the
// system stop_detection endpoint.
func EncodeStopDetectionResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*system.SystemStatus)
		enc := encoder(ctx, w)
		body := NewStopDetectionResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// marshalSystemCameraInfoToCameraInfoResponseBody builds a value of type
// *CameraInfoResponseBody from a value of type *system.CameraInfo.
func marshalSystemCameraInfoToCameraInfoResponseBody(v *system.CameraInfo) *CameraInfoResponseBody {
	res := &CameraInfoResponseBody{
		ID:            v.ID,
		Name:          v.Name,
		Device:        v.Device,
		Status:        v.Status,
		Resolution:    v.Resolution,
		Fps:           v.Fps,
		CreatedAt:     v.CreatedAt,
		AlertsEnabled: v.AlertsEnabled,
	}

	return res
}